+++ mainsequence/client/__init__.py
@@ -1,3 +1,4 @@
+from mainsequence.client.markets.models import *
 from mainsequence.logconf import logger
 
 from .command_center import *
@@ -5,5 +6,4 @@ from .models_helpers import *
 from .models_simple_tables import *
 from .models_tdag import *
 from .models_user import *
-from .models_vam import *
 from .utils import MARKETS_CONSTANTS, TDAG_CONSTANTS, AuthLoaders, bios_uuid